// src/Momentum.Frontend/apps/code-innovation-backlog/src/generated/models/Cycai_linksModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cycai_linkscycai_approvalstate = {
  100000000: 'Pending',
  100000001: 'Approved',
  100000002: 'Rejected'
} as const;
export type Cycai_linkscycai_approvalstate = keyof typeof Cycai_linkscycai_approvalstate;
export const Cycai_linksstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cycai_linksstatecode = keyof typeof Cycai_linksstatecode;
export const Cycai_linksstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cycai_linksstatuscode = keyof typeof Cycai_linksstatuscode;

export interface Cycai_linksBase {
  cycai_approvalstate: Cycai_linkscycai_approvalstate;
  "cycai_decidedbyid@odata.bind"?: string;
  cycai_decidedon?: string;
  cycai_ideaid: number;
  cycai_linkid: string;
  cycai_linkkey: string;
  cycai_name: string;
  "cycai_proposedbyid@odata.bind"?: string;
  cycai_rationale?: string;
  cycai_solutionid: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cycai_linksstatecode;
  statuscode?: Cycai_linksstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cycai_links extends Cycai_linksBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  cycai_approvalstatename?: string;
  cycai_decidedbyidname?: string;
  cycai_decidedbyidyominame?: string;
  cycai_proposedbyidname?: string;
  cycai_proposedbyidyominame?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  cycai_decidedbyid?: object;
  _cycai_decidedbyid_value?: string;
  cycai_proposedbyid?: object;
  _cycai_proposedbyid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}